+++ src/main/java/io/github/xiefrish2021/exception/NBTWriteException.java
@@ -1,7 +1,0 @@
-package io.github.xiefrish2021.exception;
-
-public class NBTWriteException extends NBTException {
-    public NBTWriteException(String message) {
-        super(message);
-    }
-}