+++ src/main/java/io/github/xiefrish2021/exception/SNBTReaderException.java
@@ -1,7 +1,0 @@
-package io.github.xiefrish2021.exception;
-
-public class SNBTReaderException extends NBTReaderException {
-    public SNBTReaderException(String message) {
-        super(message);
-    }
-}