+++ src/main/java/io/github/xiefrish2021/tag/ShortTag.java
@@ -2,7 +2,7 @@ package io.github.xiefrish2021.tag;
 
 import io.github.xiefrish2021.ITag;
 import io.github.xiefrish2021.TagType;
-import io.github.xiefrish2021.exception.NBTWriteException;
+import io.github.xiefrish2021.NBTException;
 
 @SuppressWarnings("all")
 public class ShortTag implements ITag {
@@ -10,7 +10,7 @@ public class ShortTag implements ITag {
 
     public ShortTag(Short value) {
         if (value == null) {
-            throw new NBTWriteException("The primitive type cannot be null.");
+            throw new NBTException("The primitive type cannot be null.");
         }
 
         this.value = value;