+++ src/main/java/io/github/xiefrish2021/tag/array/IntArrayTag.java
@@ -1,12 +1,64 @@
 package io.github.xiefrish2021.tag.array;
 
+import io.github.xiefrish2021.ITag;
+import io.github.xiefrish2021.NBTException;
 import io.github.xiefrish2021.TagType;
 import io.github.xiefrish2021.util.CommonUtil;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Iterator;
+import java.util.function.Consumer;
 
 @SuppressWarnings("all")
-public class IntArrayTag extends ArrayTag<Integer> {
+public class IntArrayTag implements ITag, Iterable<Integer> {
+    private final int[] array;
+
     public IntArrayTag(int[] array) {
-        super(CommonUtil.toPrimitive(array));
+        if (array == null) {
+            throw new NBTException("The array type cannot be null.");
+        }
+
+        this.array = array;
+    }
+
+    public IntArrayTag(int size) {
+        this(new int[size]);
+    }
+
+    public int size() {
+        return array.length;
+    }
+
+    @Override
+    public @NotNull Iterator<Integer> iterator() {
+        return new Iterator<>() {
+            private int index = 0;
+
+            @Override
+            public boolean hasNext() {
+                return array.length > index;
+            }
+
+            @Override
+            public Integer next() {
+                return array[index++];
+            }
+        };
+    }
+
+    public int get(int index) {
+        return array[index];
+    }
+
+    public IntArrayTag set(int index, byte value) {
+        array[index] = value;
+        return this;
+    }
+
+    public void forEach(Consumer<? super Integer> action) {
+        for (int v : array) {
+            action.accept(v);
+        }
     }
 
     @Override