+++ src/main/java/io/github/xiefrish2021/tag/array/LongArrayTag.java
@@ -1,12 +1,63 @@
 package io.github.xiefrish2021.tag.array;
 
+import io.github.xiefrish2021.ITag;
+import io.github.xiefrish2021.NBTException;
 import io.github.xiefrish2021.TagType;
-import io.github.xiefrish2021.util.CommonUtil;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Iterator;
+import java.util.function.Consumer;
 
 @SuppressWarnings("all")
-public class LongArrayTag extends ArrayTag<Long> {
+public class LongArrayTag implements ITag, Iterable<Long> {
+    private final long[] array;
+
     public LongArrayTag(long[] array) {
-        super(CommonUtil.toPrimitive(array));
+        if (array == null) {
+            throw new NBTException("The array type cannot be null.");
+        }
+
+        this.array = array;
+    }
+
+    public LongArrayTag(int size) {
+        this(new long[size]);
+    }
+
+    public int size() {
+        return array.length;
+    }
+
+    @Override
+    public @NotNull Iterator<Long> iterator() {
+        return new Iterator<>() {
+            private int index = 0;
+
+            @Override
+            public boolean hasNext() {
+                return array.length > index;
+            }
+
+            @Override
+            public Long next() {
+                return array[index++];
+            }
+        };
+    }
+
+    public long get(int index) {
+        return array[index];
+    }
+
+    public LongArrayTag set(int index, byte value) {
+        array[index] = value;
+        return this;
+    }
+
+    public void forEach(Consumer<? super Long> action) {
+        for (long v : array) {
+            action.accept(v);
+        }
     }
 
     @Override