+++ src/test/kotlin/me/coderfrish/test/KotlinTest.kt
@@ -1,6 +1,7 @@
 package me.coderfrish.test
 
 import io.github.xiefrish2021.NBT
+import io.github.xiefrish2021.tag.array.ByteArrayTag
 import io.github.xiefrish2021.tag.compound.NBTElement
 import java.io.FileInputStream
 
@@ -11,10 +12,12 @@ fun main() {
 //    compound.put("hello", StringTag("Frish2021"))
 //    nbt.writeUnnamedNBT(compound, FileOutputStream("D:\\NBT\\src\\test\\resources\\test1.nbt"))
 
-    val compound = nbt.readUnnamedNBT(FileInputStream("D:\\NBT\\src\\test\\resources\\test1.nbt"))
+    val compound = nbt.readUnnamedNBT(FileInputStream("D:\\NBT\\src\\test\\resources\\test.nbt"))
 //    println(compound["hello"].asString())
 //    val hello: NBTElement by compound
 //    println(hello.asString)
-    println(compound["hello"].asString)
+//    println(compound["hello"].asString)
 //    println(hello)
+    val sa: NBTElement by compound
+    println(sa.asIntArray)
 }